requirements.txt
reduct-py~=1.4
